refactor: Use previous instead v2 for our old encoding algo
