+++ rust/lance-encoding/src/lib.rs
@@ -16,12 +16,12 @@ pub mod decoder;
 pub mod encoder;
 pub mod encodings;
 pub mod format;
+pub mod previous;
 pub mod repdef;
 pub mod statistics;
 #[cfg(test)]
 pub mod testing;
 pub mod utils;
-pub mod v2;
 pub mod version;
 
 // We can definitely add support for big-endian machines someday.  However, it's not a priority and
