+++ rust/lance-encoding/src/previous.rs
