+++ rust/lance-encoding/src/previous/encodings/logical.rs
