+++ rust/lance-encoding/src/previous/encodings/physical/basic.rs
@@ -11,7 +11,7 @@ use crate::{
     data::{AllNullDataBlock, BlockInfo, DataBlock, NullableDataBlock},
     decoder::{PageScheduler, PrimitivePageDecoder},
     format::ProtobufUtils,
-    v2::encoder::{ArrayEncoder, EncodedArray},
+    previous::encoder::{ArrayEncoder, EncodedArray},
     EncodingsIo,
 };
