+++ rust/lance-encoding/src/previous/encodings/physical/binary.rs
@@ -20,9 +20,9 @@ use crate::encodings::physical::block::{
     BufferCompressor, CompressionConfig, GeneralBufferCompressor,
 };
 use crate::format::ProtobufUtils;
-use crate::v2::decoder::LogicalPageDecoder;
-use crate::v2::encoder::{ArrayEncoder, EncodedArray};
-use crate::v2::encodings::logical::primitive::PrimitiveFieldDecoder;
+use crate::previous::decoder::LogicalPageDecoder;
+use crate::previous::encoder::{ArrayEncoder, EncodedArray};
+use crate::previous::encodings::logical::primitive::PrimitiveFieldDecoder;
 use crate::{
     decoder::{PageScheduler, PrimitivePageDecoder},
     EncodingsIo,
